+++ src/viewpoint-measures/Unstability.cpp
@@ -22,7 +22,7 @@ void Unstability::Compute(const SceneInformationBuilder *pSceneInformationBuilde
     QVector< QVector< int > > viewpointNeighbours = pSceneInformationBuilder->GetViewpointNeighbours();
     QVector< int > viewpointsOutOfDomain;
     float minValue = FLT_MAX;
-    for( int currentViewpoint = 0; currentViewpoint < numberOfViewpoints; currentViewpoint++ )
+    for( size_t currentViewpoint = 0; currentViewpoint < numberOfViewpoints; currentViewpoint++ )
     {
         unsigned int numberOfNeighbours = 0;
         bool ocludedNeighbours = false;
@@ -61,7 +61,7 @@ void Unstability::Compute(const SceneInformationBuilder *pSceneInformationBuilde
         }
     }
     //The minimum value is assigned to the viewpoints out of the domain (viewpoints that don't see anything)
-    for( int currentViewpoint = 0; currentViewpoint < viewpointsOutOfDomain.size(); currentViewpoint++ )
+    for( size_t currentViewpoint = 0; currentViewpoint < viewpointsOutOfDomain.size(); currentViewpoint++ )
     {
         mValues[viewpointsOutOfDomain.at(currentViewpoint)] = minValue;
     }
